requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
streamlit
google-generativeai>=0.8.3
scikit-learn
pandas
pillow
pypdf
sentence-transformers
textblob